# Wedding/Wedding.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wedding.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Payment/customer/balance": {
      "get": {
        "tags": [
          "Payment"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Payment/transaction": {
      "get": {
        "tags": [
          "Payment"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterOn",
            "type": "string"
          },
          {
            "in": "query",
            "name": "filterQuery",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "isAscending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Payment/stripe/account": {
      "post": {
        "tags": [
          "Payment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStripeConnectedAccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Payment/stripe/transfer": {
      "post": {
        "tags": [
          "Payment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStripeTransferDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Payment/stripe/payout": {
      "post": {
        "tags": [
          "Payment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStripePayoutDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Payment/stripe/card": {
      "post": {
        "tags": [
          "Payment"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddStripeCardDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Website/company": {
      "get": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Website"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCompanyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Website/privacy": {
      "get": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Website"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreatePrivacyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Website"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdatePrivacyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Website/privacy/{id}": {
      "get": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Website/term-of-use": {
      "get": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Website"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTermOfUseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Website"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTermOfUseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
    "/api/Website/term-of-use/{id}": {
      "get": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Website"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddStripeCardDTO": {
      "type": "object",
      "properties": {
        "connectedAccountId": {
          "type": "string"
        },
        "cardToken": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CreatePrivacyDTO": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "CreateStripeConnectedAccountDTO": {
      "type": "object",
      "properties": {
        "refreshUrl": {
          "type": "string"
        },
        "returnUrl": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateStripePayoutDTO": {
      "type": "object",
      "properties": {
        "amount": {
          "format": "int64",
          "type": "integer"
        },
        "currency": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateStripeTransferDTO": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "amount": {
          "format": "int64",
          "type": "integer"
        },
        "currency": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateTermOfUseDTO": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "ResponseDTO": {
      "type": "object",
      "properties": {
        "result": { },
        "isSuccess": {
          "type": "boolean"
        },
        "statusCode": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCompanyDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "foundedDate": {
          "format": "date-time",
          "type": "string"
        },
        "logoUrl": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdatePrivacyDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "UpdateTermOfUseDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter your token with this format: \"Bearer YOUR_TOKEN\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}